requirements.txt
setuptools==50.3.0
python-docx==0.8.11
pytest==7.2.2
pyyaml==5.3.1
Flask==2.3.3
